[SUS-37] Refactor
