+++ src/core/util/Token.ts
